refactor: convert visits function to data source
